Refactored running using the runners
