+++ src/dotnet-test-mspec/Execution/IRunBuilder.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using System.Reflection;
+using Machine.Specifications.Runner.DotNet.Controller.Model;
+
+namespace Machine.Specifications.Runner.DotNet.Execution
+{
+    public interface IRunBuilder
+    {
+        IEnumerable<ICanRunSpecifications> BuildFrom(CommandLine commandLine, IEnumerable<Assembly> assemblies, IEnumerable<ContextInfo> contexts);
+    }
+}
